+++ mathics/builtin/graphics.py
@@ -3224,7 +3224,7 @@ clip(%s);
         )
 
         return (
-            '<mglyph width="%dpx" height="%dpx" src="data:image/svg+xml;base64,%s"/>'
+            '<mtext><img width="%dpx" height="%dpx" src="data:image/svg+xml;base64,%s"/></mtext>'
             % (
                 int(width),
                 int(height),